ai_workspace/pyproject.toml
[project]
name = "ai_workspace"
version = "0.1.0"
description = ""
authors = [
    {name = "Luciano Bermudez",email = "lberm007@ucr.edu"}
]

readme = "README.md"
requires-python =">=3.13,<3.15"
langgraph_project = { path = "ai_workspace\\langgraph_project", develop = true }
dependencies = [
    "fastapi (>=0.121.3,<0.122.0)",
    "langchain-openai (>=1.0.3,<2.0.0)",
    "pymupdf (>=1.26.6,<2.0.0)",
    "langchain (>=1.0.8,<2.0.0)",
    "langchain-core (>=1.1.0,<2.0.0)",
    "langgraph (>=1.0.3,<2.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "poetry-plugin-shell (>=1.0.1,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "langgraph-cli[inmem] (>=0.4.7,<0.5.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "faiss-cpu (>=1.13.0,<2.0.0)",
    "grpcio (>=1.76.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "grpcio-tools (>=1.76.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
